Fixed some of the cube rendering
The cube was supposed to render when you were looking at a surface you could grapple to; it would become frozen in place if you weren't looking at anything you could grapple onto. If the surface is too far away, it now turns gray, and if you aren't looking at a surface, it disappears. We can give this one a play test as well.