Added Lever-Controlled Doors, Hatches
HatchController and DoorOpen are to be used for any doors or hatches that are activated by a Reporter.

To use a Lever-Controlled Door or hatch, replace these scripts in the Door or Hatch with HatchStates or DoorStates, and in the Lever, use LeverDoor or LeverHatch as the script instead of Lever. Then, assign the Hatch or Door you want to control to the LeverDoor/LeverHatch script.

It operates as a state machine.