Added MusicReportee and functionality for track switching
Usage is as follows:
Each MusicReportee should be paired with a trigger zone that has an AreaTrigger component. The AreaTrigger now needs a MusicReportee and the Music Generator attached to it.

Each possible track for the MusicReportee should be placed in the ClipsToPlay array, and the Keys array will contain the keys for these elements. The keys are formatted as such: If you want track B to play when the second of 3 objects is fixed but not the other two, B would be in the same index in its array as the string "010" in the Keys array.

Testing of this script should follow the test cases Vittorio and I discussed in our many conversations throughout the development of this script.

Please reach out to me if you need any explanations or quick changes.